+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 - Remove duplicate unit manifests [#5]
 - Increase setuptools version [#6]
+- Decommission package [#34]
 
 0.1.0 (2022-10-11)
 ------------------